Better solution to `disable_multisource` migration
We have several source priority bug fixes in master that we have never
released because they need the `disable_multisource` setting toggled, and
that will only happen by default in Bundler 3.

The reason for this is that this setting drops support for multiple
global sources in the `gems.rb` file, and also enables separate sections
for each rubygems source in the lockfile. We have considered this to be
backwards incompatible, because `gems.rb` files using multiple global
sources have not even yet got a deprecation warning telling that they
should be changed, so we need to do that before dropping support.

So, as a compromise solution between fixing the issues that
`disable_multisource` fixes, and providing a friendly migration path for
the removal of multiple global sources support, I thought of the
following solution:

* If the `gems.rb` file has multiple top level sources, print a
deprecation message and run in the old mode (`disable_multisource` set
to false).

* If the `gems.rb` doesn't have multiple top level sources, switch
`disable_multisource` to true automatically and run in that mode.

I think this allows us to ship the most secure mode in the most common
case (no multiple top level sources), while still provide a good
migration path in the other case.

This might still break some situations where the old mode used to
resolve and the new mode does not, but it's for the sake of fixing bugs
and being more secure, so I think it's the price to pay.